move MacOS specific tests to their own directory
